Fix `AxisTest#submitInvalidAxisValue` for PCT / plugin BOM (#184)
https://github.com/jenkinsci/bom/issues/615 describes the problem in
more detail.

When the git plugin is loaded along with the matrix project plugin,
there are multiple input fields on the job configuration page with the
name `_.name`.  The previous code waited until there was at least one
input field with the name `_.name`, but instead it needs to wait for
the arrival of an additional input field with the name `_.name`.

Adds a new `setName` method that sets the value of the `_.name` field
that was created by this plugin.  Previously, it would set the first
`_.name` field.  Setting the first `_.name` field works when running
the matrix plugin tests without the git plugin loaded, but fails when
the git plugin is loaded.

The combination of waiting for the correct `_.name` field to appear and
setting the value on the correct `_.name` field allows oone of the four
tests to pass when the git plugin is loaded.  The remaining tests need
more investigation to pass when the git plugin is loaded.

Also removes a nearly silent skip of the tests when the input does not
appear within the timeout.

Also increases the time between retries so that my fast computer needs
only two or three retries.  I confirmed on a much slower computer
(Intel Core i5-2410M CPU @ 2.30GHz) that even on that slow computer,
it found the necessary field in 8 tries or less.  The upper bound of 18
tries should be more than enough for all the test environments.

The retry period should not be increased beyond the current ~300ms
because one of the tests is using a form that does not include `_.name`.
That test falls through to the end of the retry period.