fix: link on icons in Configuration matrix table
In a matrix project/build with more than one axis, the Configuration
Matrix is presented in a table where cells only contain the status icon.
The previous change which moves the icon out of the <a href> makes it
impossible to click on a cell to get to its build page. Restore the
previous behavior to make the cells clickable and allow the use of the
pop-up menu.

Signed-off-by: Michael Jeanson <mjeanson@efficios.com>