Reimplement the UI for `LabelAxis` (#213)
The `LabelAxis` configuration was using YUI TreeView. As YUI is very old and
deprecated we should remove this. This generates the complete configure page in
Jelly and just add some small JavasCript to show/hide the lists with labels.

Show the description as tooltip instead of adding it in brackets after the
label name. The description can contain HTML, so we need to safeguard it with
the markup formatter.